pyproject.toml
[tool.poetry]
name = "telegram-plubot"
version = "0.1.0"
description = ""
authors = ["apkawa <apkawa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pluggy = "^0.13.1"
python-telegram-bot = "^12.5.1"
python-meme-generator = {git="git@gitlab.com:apkawa/python-meme-generator.git"}

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-mypy = "^0.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
